Move test rule to helpers pkg.
