+++ app/src/test/java/org/mozilla/fenix/crashes/CrashReporterControllerTest.kt
@@ -14,7 +14,7 @@ import mozilla.components.support.test.ext.joinBlocking
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.mozilla.fenix.DisableNavGraphProviderAssertionRule
+import org.mozilla.fenix.helpers.DisableNavGraphProviderAssertionRule
 import org.mozilla.fenix.R
 import org.mozilla.fenix.components.Components
 import org.mozilla.fenix.components.metrics.Event
