+++ app/src/test/java/org/mozilla/fenix/helpers/DisableNavGraphProviderAssertionRule.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix
+package org.mozilla.fenix.helpers
 
 import io.mockk.every
 import io.mockk.mockkObject
